agents/pyproject.toml
[project]
name = "kisan-ai-agents"
version = "0.1.0"
description = "AI agents for Kisan platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-aiplatform[agent_engines,adk,langchain,ag2,llama_index]==1.88.0",
    "google-genai>=1.9.0",
    "google-adk==1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic==2.10.6",
    "langchain>=0.1.0",
    "langgraph>=0.1.0",
    "aiohttp>=3.12.14",
    "PyPDF2>=3.0.0",
    "python-docx>=1.1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py313"
exclude = [
    ".venv",
    "venv", 
    "__pycache__",
    "app/ai/prompts/*",
    "app/ai/prompts/**/*.py",
    "tests/*",
    "tests/**/*.py",
    "scripts/*",
    "scripts/**/*.py",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings  
    "F",  # Pyflakes
    "I",  # isort (import sorting)
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = ["B008", "B904"]
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
force-wrap-aliases = true
combine-as-imports = true
